ton-nft-lending/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
sseclient-py
tonsdk
tvm-valuetypes
aiohttp
